split commands into submodules
