+++ exercism_test_helper/lib/exercism_test_helper/cli.ex
@@ -1,6 +1,8 @@
 defmodule ExercismTestHelper.CLI do
   @moduledoc false
 
+  alias ExercismTestHelper.CLI.Command.{TestTransform, OutputLogToJSON, CombineJSON}
+
   @usage """
   Usage:
   > exercism_test_helper [--transform <filename> [--replace]]
@@ -23,89 +25,17 @@ defmodule ExercismTestHelper.CLI do
 
     cond do
       argv[:transform] ->
-        transform_test(argv[:transform], argv[:replace])
+        TestTransform.run(argv[:transform], argv[:replace])
 
       argv[:log_to_json] ->
-        log_to_json(argv[:log_to_json])
+        OutputLogToJSON.run(argv[:log_to_json])
 
       argv[:combine] ->
         [result, log] = argv[:combine] |> String.split(":")
-
-        combine(result, log)
+        CombineJSON.run(result, log)
 
       true ->
         IO.puts(@usage)
-    end
-  end
-
-  def transform_test(file, replace) do
-    output_file =
-      if replace do
-        file
-      else
-        path = Path.dirname(file)
-        name = Path.basename(file, ".exs")
-        Path.join(path, name <> "_transformed.exs")
-      end
-
-    transformed =
-      file
-      |> File.read!()
-      |> TestSource.Transformer.transform_test()
-
-    File.write!(output_file, transformed)
-  end
-
-  def log_to_json(file) do
-    json =
-      file
-      |> File.read!()
-      |> String.split("[test started] ", trim: true)
-      |> Enum.map(&String.split(&1, "\n", trim: true))
-      |> Enum.map(fn [n | os] -> {n, os |> Enum.join("\n")} end)
-      |> Enum.into(%{})
-      |> Jason.encode!()
-
-    File.write!(file <> ".json", json)
-  end
-
-  def combine(result_json_file, log_json_file) do
-    results =
-      result_json_file
-      |> File.read!()
-      |> Jason.decode!()
-
-    log =
-      log_json_file
-      |> File.read!()
-      |> Jason.decode!()
-
-    updated =
-      results["tests"]
-      |> Enum.map(&add_test_output(&1, log))
-
-    updated_results =
-      %{results | "tests" => updated}
-      |> Jason.encode!()
-
-    File.write!(result_json_file, updated_results)
-  end
-
-  defp add_test_output(%{"name" => name} = test, log) do
-    cond do
-      log[name] ->
-        output = log[name]
-
-        output =
-          cond do
-            String.length(output) > 500 -> String.slice(output, -3..-1) <> "..."
-            true -> output
-          end
-
-        Map.put(test, "output", output)
-
-      true ->
-        test
     end
   end
 end
